chore(hog-charts): rename pie-canvas-renderer to radial-canvas-renderer
Matches the rest of the radial layer (`RadialChart`, `radial-context`, `useRadialInteraction`) and leaves room for future radial chart types (donut variants, radar, sunburst) to share the file.

Generated-By: PostHog Code
Task-Id: 014c6695-6ee4-44fc-9f38-34018338582e
