+++ frontend/src/lib/hog-charts/core/radial-canvas-renderer.ts
